Complete documentation and code cleanup for shadcn/ui migration
## Documentation Updates
- Enhanced wizard README with shadcn/ui integration details
- Created comprehensive WizardFormComponents documentation
- Added accessibility features guide and migration examples
- Documented React Hook Form integration patterns

## Code Cleanup
- Removed legacy FormComponents.tsx file
- Updated forms/index.ts to export new shadcn/ui components
- Replaced old BasicInfoStep with new shadcn/ui implementation
- Exported all TypeScript interfaces for proper type safety
- Cleaned up wizard/index.ts exports (no legacy components)

## Component Consolidation
- Unified form component system under shadcn/ui
- Consistent naming and export structure
- Proper TypeScript interface exports
- Streamlined component architecture

All components now use shadcn/ui with enhanced accessibility,
proper React Hook Form integration, and comprehensive documentation.
